requirements.txt
Flask==3.0.3
pyjwt[crypto]>=2.10.1
bcrypt==4.2.0
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.32.3